ready to push

+++ auctions/views.py
@@ -8,6 +8,8 @@ from decimal import Decimal
 from .models import User, Auction, Bid, Comment
 
 
+#-------------------------- functions for module -------------------------
+
 def commoncategories(auctions):
     auctions = auctions
     categories = []
@@ -18,6 +20,7 @@ def commoncategories(auctions):
             categories.append(auction.category)
     categories.sort()
     return categories
+
 
 def commonbids(auctions):
     auctions = auctions
@@ -32,6 +35,7 @@ def commonbids(auctions):
             auctions_bids.update({auction.id : bids_for_auction[len(bids_for_auction) - 1]})
     return auctions_bids
 
+
 def actualbid(auction):
     bids = Bid.objects.filter(for_auction=auction.id)
     if bids:
@@ -42,42 +46,48 @@ def actualbid(auction):
         who_bid = 'none__$&$&$&$@@&&never--$$$$$$$4nobody__ever$$$$$$@@@@@@@@@@&&&&&&&&&'
     return bid, who_bid
 
-def index(request):
-    auctions = Auction.objects.filter(is_open='True')
-    categories = commoncategories(auctions)
-    auctions_bids = commonbids(auctions)
-    message = 'Active Listings'
-    message1 = 'No active Listings'
-    return render(request, "auctions/index.html", {
-        "auctions": auctions,
-        "auctions_bids": auctions_bids,
-        "categories": categories,
-        "message": message,
-        "message1": message1
-        })
 
-@login_required(login_url = "http://localhost:8000/login")
-def listall(request):
-    auctions = Auction.objects.all()
-    categories = commoncategories(auctions)
-    auctions_bids = commonbids(auctions)
-    message = 'All Listings'
-    return render(request, "auctions/index.html", {
-        "auctions": auctions,
-        "auctions_bids": auctions_bids,
-        "categories": categories,
-        "message": message
-        })
+def watchlist(auction, current_user):
+    follow = False
+    followers = auction.follower.all()
+    for follower in followers:
+            if follower == current_user:
+                follow = True
+    return follow    
 
+
+def new_bid(auction, bids, nbid, new, current_user):
+    message = ''
+    follow = False
+    try:
+        nbid.bid = Decimal(new)
+    except:
+        nbid.bid = 0
+    if nbid.bid < auction.init_bid and not bids:
+        message = 'Your bid must equal or greater than base bid'
+    else:
+        if bids and nbid.bid <= bids[len(bids) - 1].bid:
+            message = 'Your bid must greater than last bid'
+        else:
+            nbid.who_bid = current_user
+            nbid.for_auction = auction
+            nbid.save()
+            auction.follower.add(current_user)   
+            follow = True     
+    return message, follow
+
+
+
+#------------------------- views -----------------------------#
+
+# user area
 
 def login_view(request):
     if request.method == "POST":
-
         # Attempt to sign user in
         username = request.POST["username"]
         password = request.POST["password"]
         user = authenticate(request, username=username, password=password)
-
         # Check if authentication successful
         if user is not None:
             login(request, user)
@@ -99,7 +109,6 @@ def register(request):
     if request.method == "POST":
         username = request.POST["username"]
         email = request.POST["email"]
-
         # Ensure password matches confirmation
         password = request.POST["password"]
         confirmation = request.POST["confirmation"]
@@ -107,7 +116,6 @@ def register(request):
             return render(request, "auctions/register.html", {
                 "message": "Passwords must match."
             })
-
         # Attempt to create new user
         try:
             user = User.objects.create_user(username, email, password)
@@ -121,16 +129,53 @@ def register(request):
     else:
         return render(request, "auctions/register.html")
 
-#@login_required(login_url = "http://localhost:8000/login")
+# listings display
+
+def index(request):
+    auctions = Auction.objects.filter(is_open='True')
+    categories = commoncategories(auctions)
+    auctions_bids = commonbids(auctions)
+    message = 'Active Listings'
+    message1 = 'No active Listings'
+    return render(request, "auctions/index.html", {
+        "auctions": auctions,
+        "auctions_bids": auctions_bids,
+        "categories": categories,
+        "message": message,
+        "message1": message1
+        })
+
+
+@login_required(login_url = "http://localhost:8000/login")
+def listall(request):
+    all_auctions = Auction.objects.all()
+    auctions = Auction.objects.filter(is_open=True)
+    message = 'All Listings'
+    if all_auctions:
+        categories = commoncategories(auctions)
+        auctions_bids = commonbids(all_auctions)
+        return render(request, "auctions/index.html", {
+            "auctions": all_auctions,
+            "auctions_bids": auctions_bids,
+            "categories": categories,
+            "message": message
+            })
+    else:
+        message1 = 'No Listings'
+        return render(request, "auctions/error.html", {
+            "message": message,
+            "message1": message1
+            })
+
+
 def allcategories(request):
-    auctions = Auction.objects.all()
+    auctions = Auction.objects.filter(is_open=True)
     categories = commoncategories(auctions)
     return render(request, "auctions/allcategories.html", {
         "categories": categories
         })
 
 
-#@login_required(login_url = "http://localhost:8000/login")
 def categories(request, category):
     active_auctions = Auction.objects.filter(is_open=True)
     categories = commoncategories(active_auctions)
@@ -149,29 +194,35 @@ def categories(request, category):
 @login_required(login_url = "http://localhost:8000/login")
 def following(request):
     all_auctions = Auction.objects.all()
-    categories = commoncategories(all_auctions)
+    cat_auctions = Auction.objects.filter(is_open=True)
+    message = 'Watchlist'
+    categories = commoncategories(cat_auctions)
     follower = request.user
     auctions = Auction.objects.filter(follower=follower.id)
-    auctions_bids = commonbids(auctions)
-    message = 'Watchlist'
-    message1 = 'No Listings for this Watchlist'
-    return render(request, "auctions/index.html", {
-        "auctions": auctions,
-        "auctions_bids": auctions_bids,
-        "categories": categories,
-        "message": message,
-        "message1": message1
-        })
+    if auctions:
+        auctions_bids = commonbids(auctions)
+        return render(request, "auctions/index.html", {
+            "auctions": auctions,
+            "auctions_bids": auctions_bids,
+            "categories": categories,
+            "message": message
+            })
+    else:
+        message1 = 'No Listings for this Watchlist'
+        return render(request, "auctions/error.html", {
+            'message': message,
+            'message1': message1
+            })
 
 
 @login_required(login_url = "http://localhost:8000/login")
 def closedauctions(request):
-    auctions = Auction.objects.all()
+    auctions = Auction.objects.filter(is_open=True)
     closauctions = Auction.objects.filter(is_open=False)
     categories = commoncategories(auctions)
     auctions_bids = commonbids(closauctions)
+    message = "Closed Auctions"
     if closauctions:
-        message = "Closed Listings"
         return render(request, "auctions/index.html", {
             "auctions": closauctions,
             "auctions_bids": auctions_bids,
@@ -179,40 +230,26 @@ def closedauctions(request):
             'message': message
             })
     else:
-        message1 = "No Listing Closed"
+        message1 = "No Auctions Closed"
         return render(request, "auctions/error.html", {
+            'message': message,
             'message1': message1
             })
 
+
 def auction(request, id):
-    eauction = Auction.objects.filter(id=id).exists()
-    if eauction:
+    message = ''
+    follow = False
+    current_user = request.user
+    exists_auction = Auction.objects.filter(id=id).exists()
+    if exists_auction:
         auction = Auction.objects.get(id=id)
-        followers = auction.follower.all()
-        follow = False
-        message = ''
-        for follower in followers:
-            if follower == request.user:
-                follow = True
+        follow = watchlist(auction, current_user)
         if request.method == "POST":
             bids = Bid.objects.filter(for_auction=id)
             nbid = Bid()
             new = request.POST['nbid']
-            try:
-                nbid.bid = Decimal(new)
-            except:
-                nbid.bid = 0
-            if nbid.bid < auction.init_bid and not bids:
-                message = 'Your bid must equal or greater than base bid'
-            else:
-                if bids and nbid.bid <= bids[len(bids) - 1].bid:
-                    message = 'Your bid must greater than last bid'
-                else:
-                    nbid.who_bid = request.user
-                    nbid.for_auction = auction
-                    nbid.save()
-                    auction.follower.add(request.user)   
-                    follow = True     
+            message, follow = new_bid(auction, bids, nbid, new, current_user)
         bid, who_bid = actualbid(auction)                    
         comments = Comment.objects.filter(for_auction=id)
         return render(request, 'auctions/auction.html', {
@@ -229,26 +266,34 @@ def auction(request, id):
             'message1': message1
         })
 
-
-def error(request):
-    return render(request, 'auctions/error.html')
-
+# changes zone
 
 @login_required(login_url = "http://localhost:8000/login")
-def followadd(request):
-    follower = request.user
-    f_id = request.GET['followes']
-    auction = Auction.objects.get(id=f_id)
-    auction.follower.add(follower)
-    return HttpResponseRedirect(reverse("auction", args=[f_id]))
-
-@login_required(login_url = "http://localhost:8000/login")
-def followremove(request):
-    follower = request.user
-    f_id = request.GET['followrem']
-    auction = Auction.objects.get(id=f_id)
-    auction.follower.remove(follower)
-    return HttpResponseRedirect(reverse("auction", args=[f_id]))
+def newauction(request):
+    if request.method == "POST":
+        auction = Auction()
+        auction.name = request.POST["name"]
+        if auction.name == '':
+            auction.name = 'Nameless'
+        auction.category = (request.POST["category"]).lower()
+        if auction.category == '':
+            auction.category = 'uncategorized'
+        auction.description = request.POST["description"]
+        if auction.description == '':
+            auction.description = 'No description'
+        try:
+            auction.init_bid = abs(Decimal(request.POST["init_bid"]))
+            if auction.init_bid == 0.00:
+                auction.init_bid = 0.01
+        except:
+            auction.init_bid = 0.01
+        auction.image = request.FILES.get("image", "null")
+        auction.offerer = request.user
+        auction.save()
+        auction.follower.add(request.user)
+        return HttpResponseRedirect(reverse("auction", args=[auction.id]))
+    else:
+        return render(request, "auctions/newauction.html")
 
 
 @login_required(login_url = "http://localhost:8000/login")
@@ -271,13 +316,33 @@ def delivered(request):
 
 @login_required(login_url = "http://localhost:8000/login")
 def deleted(request):
-    status_id = request.GET['del']
-    auction = Auction.objects.get(id=status_id)
-    auction.delete()
-    message1 = 'The listing was successfully deleted'
+    if request.method == "GET":
+        status_id = request.GET['del']
+        auction = Auction.objects.get(id=status_id)
+        auction.delete()
+        message1 = 'The listing was successfully deleted'
+    else:
+        message1 = 'Page not found'
     return render(request, 'auctions/error.html', {
         'message1': message1
     })
+
+
+@login_required(login_url = "http://localhost:8000/login")
+def followadd(request):
+    follower = request.user
+    f_id = request.GET['followes']
+    auction = Auction.objects.get(id=f_id)
+    auction.follower.add(follower)
+    return HttpResponseRedirect(reverse("auction", args=[f_id]))
+
+@login_required(login_url = "http://localhost:8000/login")
+def followremove(request):
+    follower = request.user
+    f_id = request.GET['followrem']
+    auction = Auction.objects.get(id=f_id)
+    auction.follower.remove(follower)
+    return HttpResponseRedirect(reverse("auction", args=[f_id]))
 
 
 @login_required(login_url = "http://localhost:8000/login")
@@ -297,18 +362,5 @@ def newcomment(request):
         return HttpResponseRedirect(reverse("auction", args=[n_id]))
 
 
-@login_required(login_url = "http://localhost:8000/login")
-def newauction(request):
-    if request.method == "POST":
-        auction = Auction()
-        auction.name = request.POST["name"]
-        auction.category = (request.POST["category"]).lower()
-        if auction.category == '':
-            auction.category = 'uncategorized'
-        auction.description = request.POST["description"]
-        auction.init_bid = Decimal(request.POST["init_bid"])
-        auction.image = request.FILES["image"]
-        auction.offerer = request.user
-        auction.save()
-        auction.follower.add(request.user)
-    return render(request, "auctions/newauction.html")
+def error(request):
+    return render(request, 'auctions/error.html')
